+++ tests/test_app/components.py
@@ -86,6 +86,17 @@ def use_location():
 
 
 @component
+def use_origin():
+    origin = django_idom.hooks.use_origin()
+    success = bool(origin)
+    return html.div(
+        {"id": "use-origin", "data-success": success},
+        f"use_origin: {origin}",
+        html.hr(),
+    )
+
+
+@component
 def django_css():
     return html.div(
         {"id": "django-css"},
@@ -167,7 +178,7 @@ def toggle_item_mutation(item: TodoItem):
 def todo_list():
     input_value, set_input_value = hooks.use_state("")
     items = use_query(get_items_query)
-    toggle_item = use_mutation(toggle_item_mutation, refetch=get_items_query)
+    toggle_item = use_mutation(toggle_item_mutation)
 
     if items.error:
         rendered_items = html.h2(f"Error when loading - {items.error}")